Add noise estimator for Libre2 patched app data
This patch adds a new noise estimator for Libre2-like data. The code is based
around the following assumptions:

- We have raw sensor data at a higher rate than our output measurements, e.g.
  one data point per minute for Libre2 compared to the single filtered value
  that we produce every 5 minutes
- We also have that filtered data, which has been smoothed to a certain degree.

The algorithm needs to be fed with all raw and filtered data points over a given
horizon (e.g. about 15 minutes). It works in a similar way to the Dexcom noise
estimator:

- A polynomial of given order is fitted to the filtered data points. The Dexcom
  data uses a quadratic polynomial, and the Libre2 code follows that decision,
  which seems sensible as it implies a constant curvature (no kinks in the curve)

- The noise is estimated by the L2 norm of the residuals of the data points
  against the predicted values of the fitted polynomial. While the Dexcom code
  calculates the residual of the filtered values (there isn't anything else), we
  can use the raw data values that we receive once per minute.

One important difference is the fact that the filtering process will always
introduce a certain amount of lag when compared to the raw input values. This
lag increases the residuals as the slope of the curve gets steeper, which is
just an artifact of the time delay of the filtered values. In order to remove
this lag, the error estimater can debias the residuals by subtracting its mean
or median value. The norm of the resulting debiased residuals is much better
suited to signaling fluctuations in the raw input data.